+++ src/main/java/com/example/finance_manager/service/IncomeService.java
@@ -1,7 +1,9 @@
 package com.example.finance_manager.service;
 
 import com.example.finance_manager.model.Income;
+import com.example.finance_manager.model.Person;
 import com.example.finance_manager.repository.IncomeRepository;
+import com.example.finance_manager.repository.PersonRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,34 +15,54 @@ import java.util.Optional;
 public class IncomeService {
 
   private IncomeRepository incomeRepository;
+  private PersonRepository personRepository;
 
-  public void createIncome(Income income) {
-
-    incomeRepository.save(income);
+  public List<Income> fetchIncomes(String email) {
+    return incomeRepository.findAllByPersonEmail(email);
   }
 
-  public void deleteIncome(Long id) {
-
-    incomeRepository.deleteById(id);
+  public void createIncome(Income income, String email) {
+    Optional<Person> personOptional = personRepository.findByEmail(email);
+    if (personOptional.isPresent()) {
+      Person person = personOptional.get();
+      income.setPerson(person);
+      incomeRepository.save(income);
+    } else {
+      throw new RuntimeException("Person with email " + email + " not found");
+    }
   }
 
-  public List<Income> fetchIncomes() {
-
-    return incomeRepository.findAll();
+  public void deleteIncome(Long id, String email) {
+    Optional<Income> incomeOptional = incomeRepository.findById(id);
+    if (incomeOptional.isPresent()) {
+      Income income = incomeOptional.get();
+      if (income.getPerson().getEmail().equals(email)) {
+        incomeRepository.deleteById(id);
+      } else {
+        throw new RuntimeException("Unauthorized action for email " + email);
+      }
+    } else {
+      throw new RuntimeException("Income with id " + id + " not found");
+    }
   }
 
-  public void updateIncome(Long id, Income updatedIncome) {
-
+  public void updateIncome(Long id, Income updatedIncome, String email) {
     Optional<Income> existingIncomeOptional = incomeRepository.findById(id);
     if (existingIncomeOptional.isPresent()) {
       Income existingIncome = existingIncomeOptional.get();
-      existingIncome.setName(updatedIncome.getName());
-      existingIncome.setDescription(updatedIncome.getDescription());
-      existingIncome.setAnnualMonthlyValue(updatedIncome.getAnnualMonthlyValue());
-      existingIncome.setStartDate(updatedIncome.getStartDate());
-      existingIncome.setTerminationDate(updatedIncome.getTerminationDate());
-      existingIncome.setInterestRate(updatedIncome.getInterestRate());
-      incomeRepository.save(existingIncome);
+      if (existingIncome.getPerson().getEmail().equals(email)) {
+        existingIncome.setName(updatedIncome.getName());
+        existingIncome.setDescription(updatedIncome.getDescription());
+        existingIncome.setAnnualMonthlyValue(updatedIncome.getAnnualMonthlyValue());
+        existingIncome.setStartDate(updatedIncome.getStartDate());
+        existingIncome.setTerminationDate(updatedIncome.getTerminationDate());
+        existingIncome.setInterestRate(updatedIncome.getInterestRate());
+        incomeRepository.save(existingIncome);
+      } else {
+        throw new RuntimeException("Unauthorized action for email " + email);
+      }
+    } else {
+      throw new RuntimeException("Income with id " + id + " not found");
     }
   }
 }
